read me
# LostAndFound
Mini Projet Andriod
This project is a mobile application project designed for everybody which provides the easiest way to make searchs for public services.
This application lets the user create an account, search for a lost item and even pay the one who provides the service online. Our application does also create the ability to add a found item and also to put all lost object withing 6 months to donations for charity.
Our application is name Lost & Found and is developped in Swift for The client and in node.js for the backend part with mongoDB as database  
<p align="center">
<img src="[https://shiva-engine.com/wp-content/uploads/2017/10/xcode-logo-small.png](https://media.techtribune.net/uploads/2022/02/2022-02-25-ts3_thumbs-8b5.png)" width="300" height="300">
<img src="https://cdn-icons-png.flaticon.com/512/38/38002.png" width="300" height="300">
</p>

<h1 align="center"><project-name></h1>

<p align="center"><project-description></p>

## Links

- [Mini Project](https://github.com/ghaithghanem/Android)
- [Backend] (https://github.com/mohamedmmc/LostAndFound-Back)
- [AndoidStudio] (https://developer.android.com/)

## Available Commands

In the project directory, you can run:

### Excute the project
### Change the storyboard wich is the UI of the app

The app is built using Android Studio. Run the backend using the link in ##Links then build the app to start using the routes of the back
In this app, you will have to create an account using facebook, google or github Authenticator
Once logged in, The user can either add a Lost or Found Item, which other users will see updated.
All of the lost and find Item have to be initilized with a position on the map, so the one searching for an item can have an idea where it has been pinged
  
### Coming Soon
The app is not finished yet, so all of the other implementations will be added later !


## Built With

- Android Studio
- Node js 
- MongoDB